Refactor: remove promises utils as public APIs
