+++ src/utils/index.ts
@@ -4,6 +4,5 @@ export { convertTransactionStructToCommitmentSummary } from './commitment';
 export * from './ecies';
 export * from './encryption/aes';
 export { getPublicViewingKey, verifyED25519 } from './keys-utils';
-export * from './promises';
 export * from './runtime';
 export * from './stringify';
