+++ src/wallet/abstract-wallet.ts
@@ -112,7 +112,7 @@ import {
   formatTXOsReceivedPOIStatusInfo,
   formatTXOsSpentPOIStatusInfo,
 } from '../poi/poi-status-formatter';
-import { delay, stringifySafe } from '../utils';
+import { stringifySafe } from '../utils';
 import {
   getRailgunTransactionIDFromBigInts,
   getRailgunTxidLeafHash,
@@ -122,6 +122,7 @@ import { POIValidation } from '../validation/poi-validation';
 import { extractFirstNoteERC20AmountMapFromTransactionRequest } from '../validation/extract-transaction-data';
 import { Registry } from '../utils/registry';
 import { CURRENT_UTXO_MERKLETREE_HISTORY_VERSION, ZERO_32_BYTE_VALUE } from '../utils/constants';
+import { delay } from '../utils/promises';
 
 type ScannedDBCommitment = PutBatch<string, Buffer>;
